Move prisma functions to lib folder
