+++ pages/api/posts.ts
@@ -1,18 +1,9 @@
 import { Prisma } from '@prisma/client'
-import prisma from 'lib/prisma'
+import { addPost, getPostCount, getPosts } from 'lib/prisma/posts'
 
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 type TPost = Prisma.PostGetPayload<true>
-
-interface IPost {
-  id: number
-  author: string
-  title: string
-  date: Date
-  content: string
-  [x: string]: unknown
-}
 
 interface NextApiRequestForPost extends NextApiRequest {
   body: TPost
@@ -21,42 +12,6 @@ interface NextApiRequestForPost extends NextApiRequest {
 interface Result {
   success: boolean
   message: string | Record<string, unknown> | Array<Record<string, unknown>>
-}
-
-export const addPost = async (post: TPost): Promise<void> => {
-  post.published = true
-  post.date = new Date()
-
-  await prisma.post.create({
-    data: post,
-  })
-}
-
-export const getPosts = async (page?: number): Promise<IPost[]> => {
-  const posts = await prisma.post.findMany({
-    select: {
-      id: true,
-      author: true,
-      title: true,
-      date: true,
-      content: true,
-    },
-    where: {
-      published: {
-        equals: true,
-      },
-    },
-    orderBy: {
-      id: 'desc',
-    },
-    skip: page !== undefined ? (page - 1) * 10 : undefined,
-    take: page !== undefined ? 10 : undefined,
-  })
-  return posts
-}
-
-export const getPostCount = async (): Promise<number> => {
-  return await prisma.post.count()
 }
 
 const Post = async (
